Move key tests to their own test suite
